COL-2: committed seeders for fee tables

+++ database/migrations/users/2020_03_26_010000_create_fees_table.php
@@ -17,7 +17,7 @@ class CreateFeesTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('fee_type_id');
             $table->string('name');
-            $table->float('percentage', 2, 2);
+            $table->smallInteger('percentage');
             $table->timestamps();
 
             $table->foreign('fee_type_id')
